requirement.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-crispy-forms==1.11.1
django-filter==2.4.0
django-six==1.0.4
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyJWT==2.0.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-crispy-forms==1.11.1
django-filter==2.4.0
django-six==1.0.4
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
MarkupSafe==1.1.1
PyJWT==2.0.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
django-tables2==2.4.0